xUnit for UserPosition controller

+++ UoW/Controllers/Users/UserPositionController.cs
@@ -45,24 +45,44 @@ namespace UoW.Controllers.Users
 			return Ok(response);
 		}
 
-		[HttpPost]
-		public async Task<IActionResult> SaveUserPosition(UserPositionRequest request)
+		[HttpPost("update")]
+		public async Task<IActionResult> Update(UserPositionRequest request)
 		{
 			if (request == null) return NotFound(request);
 
 			var position = _mapper.Map<UserPosition>(request);
 
-			var result = await _userPositionService.SaveUserPosition(position);
+			var result = await _userPositionService.Update(position);
 
-			return Ok(result);
+			var response = _mapper.Map<UserPositionResponse>(result);
+
+			return Ok(response);
+		}
+
+		[HttpPost("create")]
+		public async Task<IActionResult> Create(UserPositionRequest request)
+		{
+			if (request == null) return BadRequest(request);
+
+			var position = _mapper.Map<UserPosition>(request);
+
+			var result = await _userPositionService.Create(position);
+
+			var response = _mapper.Map<UserPositionResponse>(result);
+
+			return Ok(response);
 		}
 
 		[HttpDelete]
 		public async Task<IActionResult> DeleteUserPosition(int positioId)
 		{
+			var userPos = await _userPositionService.GetUserPosition(positioId);
+
+			if (userPos == null) return NotFound(positioId);
+
 			await _userPositionService.DeleteUserPosition(positioId);
 
-			return Ok();
+			return Ok(positioId);
 		}
 
 	}
